requirements.txt
geopandas
seaborn
psycopg2
geoalchemy2
xlrd
xlsxwriter
